restructure api client module
