+++ commons/src/main/java/com/services/commons/Application.java
@@ -1,13 +1,13 @@
-package com.services.fetchrestaurantdata;
+package com.services.commons;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class FetchrestaurantdataApplication {
+public class Application {
 
     public static void main(String[] args) {
-        SpringApplication.run(FetchrestaurantdataApplication.class, args);
+        SpringApplication.run(Application.class, args);
     }
 
 }
