--- commons/src/test/java/com/services/commons/CommonsApplicationTests.java
@@ -1,13 +1,0 @@
-package com.services.commons;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class CommonsApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
