--- persist-restaurant-data/src/test/java/com/services/persistrestaurantdata/PersistRestaurantDataApplicationTests.java
@@ -1,13 +1,0 @@
-package com.services.persistrestaurantdata;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class PersistRestaurantDataApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
