+++ restaurant-data-api-client/src/main/java/com/services/fetchrestaurantdata/Application.java
@@ -1,13 +1,13 @@
-package com.services.commons;
+package com.services.fetchrestaurantdata;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class CommonsApplication {
+public class Application {
 
     public static void main(String[] args) {
-        SpringApplication.run(CommonsApplication.class, args);
+        SpringApplication.run(Application.class, args);
     }
 
 }
