--- restaurant-data-api-client/src/test/java/com/services/fetchrestaurantdata/FetchrestaurantdataApplicationTests.java
@@ -1,13 +1,0 @@
-package com.services.fetchrestaurantdata;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class FetchrestaurantdataApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
